requirements.txt
diffusers==0.28.2
transformers==4.41.2
accelerate==0.30.1
boto3==1.35.18
hf-transfer==0.1.9
huggingface-hub==0.27.1
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
zipp==3.16.2
Pillow==10.0.0
pydantic==2.1.1
pydantic_core==2.4.0
fastapi==0.101.0